Use common resource hook for roles, extended with error
